requirements.txt
pandas
openpyxl
openai
boto3
ddgs
requests
beautifulsoup4
tabulate
python-dotenv
